third_party/blink/web_tests/platform/mac/external/wpt/mathml/presentation-markup/mpadded/mpadded-002-expected.txt

This is a testharness.js-based test.
FAIL mpadded with no attributes assert_approx_equals: content height expected 150 +/- 1 but got -4
FAIL Different widths assert_approx_equals: width 0 expected 25 +/- 1 but got 100
FAIL Different heights assert_approx_equals: height0 expected 25 +/- 1 but got -4
FAIL Percentage calculation for width, height and depth assert_equals: width expected 50 but got 100
FAIL Different depths assert_approx_equals: depth0 expected 25 +/- 1 but got 125
FAIL Various combinations of height, depth and width. assert_approx_equals: width0 expected 25 +/- 1 but got 100
PASS Preferred width
PASS dynamic attributes (remove)
FAIL dynamic attributes (attach) assert_approx_equals: expected 50 +/- 1 but got 100
FAIL dynamic attributes (modify) assert_approx_equals: expected 50 +/- 1 but got 100
Harness: the test ran to completion.
